refactor(services): split client.go user methods into client_user.go
- Extract UserUpdateClientRequest struct into client_user.go
- Extract validateUserScopes, ListClientsByUser into client_user.go
- Extract UserUpdateClient, UserDeleteClient into client_user.go
- Keep admin methods, shared types, and helpers in client.go
